Require the board name to be typed before deleting

+++ client/js/controllers/project.js
@@ -8,11 +8,25 @@ App.ProjectsIndexRoute = Ember.Route.extend({
 
 App.ProjectsNewRoute = Ember.Route.extend({
   redirect: function () {
-    this.transitionTo('projects.edit', this.store.createRecord('project', { name: 'New' }).save());
+    this.transitionTo('projects.edit', this.store.createRecord('project', {
+      name: 'New' // + moment().format('YYYYMMDDHHmmss')
+    }).save());
+  }
+});
+
+App.ProjectsEditRoute = Ember.Route.extend({
+  setupController: function (controller, project) {
+    controller.setProperties({
+      confirmDelete: null,
+      content: project
+    });
   }
 });
 
 App.ProjectsEditController = Ember.ObjectController.extend({
+
+  confirmDelete: null,
+
   actions: {
     save: function () {
       var project = this.get('content');
@@ -21,9 +35,11 @@ App.ProjectsEditController = Ember.ObjectController.extend({
     },
     'delete': function () {
       var project = this.get('content');
-      project.deleteRecord();
-      project.save();
-      this.transitionToRoute('projects');
+      if (this.get('confirmDelete') === project.get('name')) {
+        project.deleteRecord();
+        project.save();
+        this.transitionToRoute('projects');
+      }
     }
   }
 });
